+++ Localization/zhCN.ExtraWrathStrings.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-09-07 07-25-47 ]] --
+-- [[ Exported at 2022-09-17 22-03-46 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --